chore: correct compiler warnings

+++ src/mrfio.cpp
@@ -79,7 +79,7 @@ void MRFExporter::export_seq(const string& seq, const size_t& width, ostream& os
 
 void MRFExporter::export_psfm(const Float1dArray& w, ostream& os) {
     export_elem(w(0), os);
-    for (int i = 1; i < w.size(); ++i) export_elem(w(i), os, true);
+    for (size_t i = 1; i < w.size(); ++i) export_elem(w(i), os, true);
 }
 
 void MRFExporter::export_node_symbol(const string& sym, ostream& os) {
@@ -89,7 +89,7 @@ void MRFExporter::export_node_symbol(const string& sym, ostream& os) {
 
 void MRFExporter::export_node_weight(const Float1dArray& w, ostream& os) {
     export_elem(w(0), os);
-    for (int i = 1; i < w.size(); ++i) export_elem(w(i), os, true);
+    for (size_t i = 1; i < w.size(); ++i) export_elem(w(i), os, true);
 }
 
 void MRFExporter::export_edge_symbol(const string& sym, ostream& os) {
@@ -239,21 +239,21 @@ void MRFImporter::import_body(istream& is, MRF& model) {
 
 Float1dArray MRFImporter::import_node_weight(istream& is, const size_t& num_var) {
     Float1dArray w(num_var);
-    for (int i = 0; i < num_var; ++i) w(i) = import_elem(is);
+    for (size_t i = 0; i < num_var; ++i) w(i) = import_elem(is);
     return w;
 }
 
 Float2dArray MRFImporter::import_edge_weight(istream& is, const size_t& num_var) {
     Float2dArray w(num_var, num_var);
-    for (int i = 0; i < num_var; ++i)
-        for (int j = 0; j < num_var; ++j)
-            w(i, j) = import_elem(is);
+    for (size_t i = 0; i < num_var; ++i)
+        for (size_t j = 0; j < num_var; ++j)
+            w((int)i, (int)j) = import_elem(is);
     return w;
 }
 
 Float1dArray MRFImporter::import_psfm(istream& is, const size_t& num_var) {
     Float1dArray w(num_var);
-    for (int i = 0; i < num_var; ++i) w(i) = import_elem(is);
+    for (size_t i = 0; i < num_var; ++i) w(i) = import_elem(is);
     return w;
 }
